+++ src/main/java/com/oltpbenchmark/benchmarks/tpcc/procedures/Payment.java
@@ -253,7 +253,8 @@ public class Payment extends TPCCProcedure {
 
     private void updateWarehouse(Connection conn, int w_id, float paymentAmount) throws SQLException {
         try (PreparedStatement payUpdateWhse = this.getPreparedStatement(conn, payUpdateWhseSQL)) {
-            payUpdateWhse.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+            // payUpdateWhse.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+            payUpdateWhse.setDouble(1, paymentAmount);
             payUpdateWhse.setInt(2, w_id);
             // MySQL reports deadlocks due to lock upgrades:
             // t1: read w_id = x; t2: update w_id = x; t1 update w_id = x
@@ -308,7 +309,8 @@ public class Payment extends TPCCProcedure {
 
     private void updateDistrict(Connection conn, int w_id, int districtID, float paymentAmount) throws SQLException {
         try (PreparedStatement payUpdateDist = this.getPreparedStatement(conn, payUpdateDistSQL)) {
-            payUpdateDist.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+            // payUpdateDist.setBigDecimal(1, BigDecimal.valueOf(paymentAmount));
+            payUpdateDist.setDouble(1, paymentAmount);
             payUpdateDist.setInt(2, w_id);
             payUpdateDist.setInt(3, districtID);
